Add files via upload
this a project that is a combination of power system and machine learning and python. Basically the project is all about the identification of low frequencies from the raw data we already have from pmus. In the three codes uploaded in one code we try to get a csv file for a particular case that has active and reactive power from the raw data that had parameters for eg. voltage magnitude ,voltage angle, current magnitude, current angle etc. for different buses. Then from this csv file we get another csv files where it has data for a single bus at all time stamps of a particular case. And from the active and reactive power files we get the final eigenvalues and dmd modes csv files. And on these final csv files we will apply dmd and mrdmd for the identification of low frequency modes and train a model that identify these low frequency modes on its own.